requirements.txt
# Core
django==4.2.9
django-redis==5.4.0
python-dotenv==1.0.1
str2bool==1.1

# Utils
Pillow==10.0.1
envparse==0.2.0

# Services
celery==5.3.4
redis==5.0.1
hiredis==2.2.3

# Tooling
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

# DB Layer
psycopg2-binary==2.9.9

firebase-admin==6.5.0
openai==1.55.3
sentry-sdk[django]
